Added plot trade and plot trends Plot trade is called from strategy and will plot the price action, indicator values, and trade entry and exit. Plot trends is called from trade collection and will show how a linear fit of current return to end trade return changes over time.

+++ src/System/TestScript.py
@@ -237,7 +237,7 @@ def test_pars(short_pars, long_pars):
 
     for long in long_pars:
         for short in short_pars:
-            strat.measure.update_param((short, long))
+            strat.measure.update_param((long, short))
             strat.refresh()
             sharpes.loc[short, long] = strat.trades.Sharpe_annual
             summaries.append(('{}-{}'.format(short, long), strat.trades.summary_report()))
